+++ src/Dynamo.All.2012.sln
@@ -77,7 +77,7 @@ Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "DynamoCoreWpfTests", "..\te
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "IronPythonTests", "..\test\Libraries\DynamoPythonTests\IronPythonTests.csproj", "{773988FE-EDF6-45CB-A63F-482955EB3553}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "IronPythonEval", "Libraries\DSIronPython\IronPythonEval.csproj", "{9EEF4F42-6B3B-4358-9A8A-C2701539A822}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "DSIronPython", "Libraries\DSIronPython\DSIronPython.csproj", "{9EEF4F42-6B3B-4358-9A8A-C2701539A822}"
 	ProjectSection(ProjectDependencies) = postProject
 		{133FC760-5699-46D9-BEA6-E816B5F01016} = {133FC760-5699-46D9-BEA6-E816B5F01016}
 	EndProjectSection
@@ -203,6 +203,8 @@ EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Display", "Libraries\Display\Display.csproj", "{A50C198C-DA6E-4081-BC53-0F44D287F207}"
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "DisplayTests", "..\test\Libraries\DisplayTests\DisplayTests.csproj", "{F8D2E3E5-4505-4463-9367-2EB2629D7DCD}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "PythonNodeModels", "Libraries\PythonNodeModels\PythonNodeModels.csproj", "{8872CA17-C10D-43B9-8393-5C5A57065EB0}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -474,6 +476,10 @@ Global
 		{F8D2E3E5-4505-4463-9367-2EB2629D7DCD}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{F8D2E3E5-4505-4463-9367-2EB2629D7DCD}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{F8D2E3E5-4505-4463-9367-2EB2629D7DCD}.Release|Any CPU.Build.0 = Release|Any CPU
+		{8872CA17-C10D-43B9-8393-5C5A57065EB0}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{8872CA17-C10D-43B9-8393-5C5A57065EB0}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{8872CA17-C10D-43B9-8393-5C5A57065EB0}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{8872CA17-C10D-43B9-8393-5C5A57065EB0}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
@@ -503,6 +509,7 @@ Global
 		{E674F1A1-BE83-475A-9CC9-F55CADBEC448} = {FA7BE306-A3B0-45FA-9D87-0C69E6932C13}
 		{C068DE76-0651-4F6C-8807-B90EF749D104} = {FA7BE306-A3B0-45FA-9D87-0C69E6932C13}
 		{A50C198C-DA6E-4081-BC53-0F44D287F207} = {FA7BE306-A3B0-45FA-9D87-0C69E6932C13}
+		{8872CA17-C10D-43B9-8393-5C5A57065EB0} = {FA7BE306-A3B0-45FA-9D87-0C69E6932C13}
 		{0355EB79-6961-4B8A-8F66-035D4EC0C9FE} = {F4D44BC0-32CF-4E58-AD2A-F19CE1450B00}
 		{6E1177DB-5E4F-4E2E-82E7-902437E56AED} = {0355EB79-6961-4B8A-8F66-035D4EC0C9FE}
 		{2235F7CA-AB25-4177-92BC-4B574D2D5DF6} = {0355EB79-6961-4B8A-8F66-035D4EC0C9FE}
