+++ src/Libraries/PythonNodeModels/Properties/AssemblyInfo.cs
@@ -5,7 +5,7 @@ using System.Runtime.InteropServices;
 // set of attributes. Change these attribute values to modify the information
 // associated with an assembly.
 
-[assembly: AssemblyTitle("PythonNodeModelsWpf")]
+[assembly: AssemblyTitle("PythonNodeModels")]
 [assembly: AssemblyCulture("")]
 // The following GUID is for the ID of the typelib if this project is exposed to COM
-[assembly: Guid("5df40db3-627c-43fb-9a0e-0ee422d92485")]
+[assembly: Guid("a0debf47-0cf2-46e8-90b9-0c3d9b808fd7")]
