+++ src/Libraries/PythonNodeModels/Properties/Resources.en-US.resx
@@ -1,4 +1,4 @@
-﻿<?xml version="1.0" encoding="utf-8"?>
+<?xml version="1.0" encoding="utf-8"?>
 <root>
   <!-- 
     Microsoft ResX Schema 
@@ -117,9 +117,6 @@
   <resheader name="writer">
     <value>System.Resources.ResXResourceWriter, System.Windows.Forms, Version=4.0.0.0, Culture=neutral, PublicKeyToken=b77a5c561934e089</value>
   </resheader>
-  <data name="EditHeader" xml:space="preserve">
-    <value>Edit...</value>
-  </data>
   <data name="PythonNodePortDataOutputToolTip" xml:space="preserve">
     <value>Result of the python script</value>
   </data>
