+++ test/DynamoCoreWpfTests/RecordedTests.cs
@@ -6,7 +6,6 @@ using System.Xml;
 
 using SystemTestServices;
 
-using DSIronPythonNode;
 using Dynamo;
 using Dynamo.Controls;
 using Dynamo.Models;
@@ -19,6 +18,8 @@ using NUnit.Framework;
 using Dynamo.UI;
 using System.Reflection;
 using TestServices;
+
+using PythonNodeModels;
 
 using IntegerSlider = DSCoreNodesUI.Input.IntegerSlider;
